Move InMemoryMetricReader to :sdk:metrics-testing, deprecate version in :sdk:metrics
