+++ sdk/trace/build.gradle.kts
@@ -41,6 +41,7 @@ dependencies {
     // dependencies.
     isTransitive = false
   }
+  jmh(project(":sdk:metrics-testing"))
   jmh(project(":exporters:jaeger-thrift"))
   jmh(project(":exporters:otlp:trace")) {
     // The opentelemetry-exporter-otlp-trace depends on this project itself. So don"t pull in
